+++ lib/sfio/sfprint.c
@@ -297,17 +297,16 @@ int sfprint(FILE *f, Sffmt_t *format) {
 	/* set the correct size */
 	if (flags & (SFFMT_TYPES & ~SFFMT_IFLAG)) {
 	    if (_Sftype[fmt] & (SFFMT_INT | SFFMT_UINT)) {
-		size = (flags & SFFMT_LLONG) ? sizeof(long long) :
-		    (flags & SFFMT_LONG) ? sizeof(long) :
-		    (flags & SFFMT_SHORT) ? sizeof(short) :
-		    (flags & SFFMT_SSHORT) ? sizeof(char) :
-		    (flags & SFFMT_JFLAG) ? sizeof(long long) :
-		    (flags & SFFMT_TFLAG) ? sizeof(ptrdiff_t) :
-		    (flags & SFFMT_ZFLAG) ? sizeof(size_t) : -1;
+		size = (flags & SFFMT_LLONG)  ? (ssize_t)sizeof(long long) :
+		       (flags & SFFMT_LONG)   ? (ssize_t)sizeof(long) :
+		       (flags & SFFMT_SHORT)  ? (ssize_t)sizeof(short) :
+		       (flags & SFFMT_SSHORT) ? (ssize_t)sizeof(char) :
+		       (flags & SFFMT_JFLAG)  ? (ssize_t)sizeof(long long) :
+		       (flags & SFFMT_TFLAG)  ? (ssize_t)sizeof(ptrdiff_t) :
+		       (flags & SFFMT_ZFLAG)  ? (ssize_t)sizeof(size_t) : -1;
 	    } else if (_Sftype[fmt] & SFFMT_FLOAT) {
-		size = (flags & SFFMT_LDOUBLE) ? sizeof(long double) :
-		    (flags & (SFFMT_LONG | SFFMT_LLONG)) ?
-		    sizeof(double) : -1;
+		size = (flags & SFFMT_LDOUBLE)              ? (ssize_t)sizeof(long double) :
+		       (flags & (SFFMT_LONG | SFFMT_LLONG)) ? (ssize_t)sizeof(double) : -1;
 	    }
 	}
 
@@ -466,9 +465,10 @@ int sfprint(FILE *f, Sffmt_t *format) {
 		    break;
 		if (lv < 0 && fmt == 'd') {
 		    flags |= SFFMT_MINUS;
-		    if (lv == HIGHBITL) {	/* avoid overflow */
-			lv = (long long)(HIGHBITL / base);
-			*--sp = _Sfdigits[HIGHBITL - (unsigned long long)lv * base];
+		    if ((unsigned long long)lv == HIGHBITL) { // avoid overflow
+			lv = (long long)(HIGHBITL / (unsigned long long)base);
+			*--sp = _Sfdigits[HIGHBITL -
+			                  (unsigned long long)lv * (unsigned long long)base];
 		    } else
 			lv = -lv;
 		}
@@ -481,8 +481,8 @@ int sfprint(FILE *f, Sffmt_t *format) {
 		    } while ((lv = (unsigned long long)lv >> n));
 		} else {	/* general base */
 		    do {
-			*--sp = ssp[(unsigned long long)lv % base];
-		    } while ((lv = (unsigned long long)lv / base));
+			*--sp = ssp[(unsigned long long)lv % (unsigned long long)base];
+		    } while ((lv = (unsigned long long)lv / (unsigned long long)base));
 		}
 	    } else
 	    if (sizeof(short) < sizeof(int) && FMTCMP(size, short, long long)) {
@@ -518,9 +518,9 @@ int sfprint(FILE *f, Sffmt_t *format) {
 		    break;
 		if (v < 0 && fmt == 'd') {
 		    flags |= SFFMT_MINUS;
-		    if (v == HIGHBITI) {	/* avoid overflow */
-			v = (int) (HIGHBITI / base);
-			*--sp = _Sfdigits[HIGHBITI - (unsigned)v * base];
+		    if ((unsigned)v == HIGHBITI) { // avoid overflow
+			v = (int)(HIGHBITI / (unsigned)base);
+			*--sp = _Sfdigits[HIGHBITI - (unsigned)v * (unsigned)base];
 		    } else
 			v = -v;
 		}
@@ -532,8 +532,8 @@ int sfprint(FILE *f, Sffmt_t *format) {
 		    } while ((v = (unsigned)v >> n));
 		} else {	/* n_s == 0, general base */
 		    do {
-			*--sp = ssp[(unsigned)v % base];
-		    } while ((v = (unsigned)v / base));
+			*--sp = ssp[(unsigned)v % (unsigned)base];
+		    } while ((v = (unsigned)v / (unsigned)base));
 		}
 	    }
